refactor(cliproxy): extract dry-run preview formatter into setup/preview.ts
Move DryRunPreviewOptions interface and formatDryRunPreview function into a
focused submodule. Adds 2 new TDD-first byte-count regression tests for the
anthropic-only (51 bytes) and dual-provider (98 bytes) OPENCODE_AUTH_JSON
output, locking in the v0.7.0 dry-run contract.

setup.ts shrinks from 555 to 518 LOC (724 tests total, +2).
