+++ packages/cli/src/commands/cliproxy/setup.test.ts
@@ -5,12 +5,12 @@ import {goke} from 'goke'
 
 import {
   buildNonInteractivePlan,
-  formatDryRunPreview,
   mustConfirmDestructive,
   registerCliproxySetup,
   validateSetupOptions,
   verifyModelsAvailable,
 } from './setup'
+import {formatDryRunPreview} from './setup/preview'
 import {getHarnessTemplate} from './setup/templates'
 
 describe('cliproxy setup helpers', () => {
@@ -253,117 +253,6 @@ describe('destructive overwrite UX', () => {
 
     it("['openai', 'anthropic'] → true (order does not matter)", () => {
       expect(mustConfirmDestructive(['openai', 'anthropic'])).toBe(true)
-    })
-  })
-
-  // ── formatDryRunPreview ───────────────────────────────────────────────────
-
-  describe('formatDryRunPreview', () => {
-    it('renders the dry-run header with repo and providers', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: KEY, baseUrl: BASE_URL})
-      const preview = formatDryRunPreview({
-        repo: 'owner/repo',
-        harness: 'opencode',
-        providers: ['anthropic'],
-        model: 'anthropic/claude-sonnet-4-6',
-        template,
-      })
-
-      expect(preview).toContain('Dry run: cliproxy setup --harness opencode')
-      expect(preview).toContain('Repository: owner/repo')
-      expect(preview).toContain('Providers: anthropic')
-    })
-
-    it('renders planned secrets with byte sizes', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: KEY, baseUrl: BASE_URL})
-      const preview = formatDryRunPreview({
-        repo: 'owner/repo',
-        harness: 'opencode',
-        providers: ['anthropic'],
-        model: 'anthropic/claude-sonnet-4-6',
-        template,
-      })
-
-      expect(preview).toContain('Planned secrets:')
-      expect(preview).toContain('OPENCODE_AUTH_JSON')
-      expect(preview).toContain('OPENCODE_CONFIG')
-      expect(preview).toContain('OMO_PROVIDERS')
-    })
-
-    it('renders planned variables', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: KEY, baseUrl: BASE_URL})
-      const preview = formatDryRunPreview({
-        repo: 'owner/repo',
-        harness: 'opencode',
-        providers: ['anthropic'],
-        model: 'anthropic/claude-sonnet-4-6',
-        template,
-      })
-
-      expect(preview).toContain('Planned variables:')
-      expect(preview).toContain('FRO_BOT_MODEL')
-    })
-
-    it('renders proxy key as <proxy-key> placeholder, NOT the actual key value', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: KEY, baseUrl: BASE_URL})
-      const preview = formatDryRunPreview({
-        repo: 'owner/repo',
-        harness: 'opencode',
-        providers: ['anthropic'],
-        model: 'anthropic/claude-sonnet-4-6',
-        template,
-      })
-
-      expect(preview).toContain('<proxy-key>')
-      expect(preview).not.toContain(KEY)
-    })
-
-    it('renders "No mutations will be performed." footer', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: KEY, baseUrl: BASE_URL})
-      const preview = formatDryRunPreview({
-        repo: 'owner/repo',
-        harness: 'opencode',
-        providers: ['anthropic'],
-        model: 'anthropic/claude-sonnet-4-6',
-        template,
-      })
-
-      expect(preview).toContain('No mutations will be performed.')
-    })
-
-    it('dual-provider preview lists both providers', () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: KEY,
-        baseUrl: BASE_URL,
-        providers: ['anthropic', 'openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const preview = formatDryRunPreview({
-        repo: 'owner/repo',
-        harness: 'opencode',
-        providers: ['anthropic', 'openai'],
-        model: 'openai/gpt-5.4-mini',
-        template,
-      })
-
-      expect(preview).toContain('anthropic')
-      expect(preview).toContain('openai')
-      expect(preview).not.toContain(KEY)
-    })
-
-    it('secret values in preview do NOT contain the actual key value', () => {
-      // Even if the template has the key embedded in JSON, the preview must redact it
-      const template = getHarnessTemplate('opencode', {keyValue: KEY, baseUrl: BASE_URL})
-      const preview = formatDryRunPreview({
-        repo: 'owner/repo',
-        harness: 'opencode',
-        providers: ['anthropic'],
-        model: 'anthropic/claude-sonnet-4-6',
-        template,
-      })
-
-      // The actual key must not appear anywhere in the preview output
-      expect(preview).not.toContain(KEY)
     })
   })
